refactor(repo C6): extract chunking to engine/repository/chunking.ts
Moves chunkDocument and chunkDocumentByTokens. Depends on chunker.js
(scanBreakPoints, findCodeFences, findBestCutoff, chunk constants)
and inference.js (getDefaultLLM for tokenization).

Build: clean. Tests: 970/970 pass.
