+++ engine/repository.ts
@@ -136,6 +136,7 @@ import {
   parsePositiveInt,
   runWithConcurrencyLimit,
 } from "./repository/rerank-queue.js";
+import { chunkDocument, chunkDocumentByTokens } from "./repository/chunking.js";
 
 // =============================================================================
 // Virtual Path Utilities (kindx://)
@@ -1299,161 +1300,7 @@ export function getActiveDocumentPaths(db: Database, collectionName: string): st
 
 export { formatQueryForEmbedding, formatDocForEmbedding };
 
-export function chunkDocument(
-  content: string,
-  maxChars: number = CHUNK_SIZE_CHARS,
-  overlapChars: number = CHUNK_OVERLAP_CHARS,
-  windowChars: number = CHUNK_WINDOW_CHARS
-): { text: string; pos: number }[] {
-  if (content.length <= maxChars) {
-    return [{ text: content, pos: 0 }];
-  }
-
-  // Pre-scan all break points and code fences once
-  const breakPoints = scanBreakPoints(content);
-  const codeFences = findCodeFences(content);
-
-  const chunks: { text: string; pos: number }[] = [];
-  let charPos = 0;
-
-  while (charPos < content.length) {
-    // Calculate target end position for this chunk
-    const targetEndPos = Math.min(charPos + maxChars, content.length);
-
-    let endPos = targetEndPos;
-
-    // If not at the end, find the best break point
-    if (endPos < content.length) {
-      // Find best cutoff using scored algorithm
-      const bestCutoff = findBestCutoff(
-        breakPoints,
-        targetEndPos,
-        windowChars,
-        0.7,
-        codeFences
-      );
-
-      // Only use the cutoff if it's within our current chunk
-      if (bestCutoff > charPos && bestCutoff <= targetEndPos) {
-        endPos = bestCutoff;
-      }
-    }
-
-    // Ensure we make progress
-    if (endPos <= charPos) {
-      endPos = Math.min(charPos + maxChars, content.length);
-    }
-
-    chunks.push({ text: content.slice(charPos, endPos), pos: charPos });
-
-    // Move forward, but overlap with previous chunk
-    // For last chunk, don't overlap (just go to the end)
-    if (endPos >= content.length) {
-      break;
-    }
-    charPos = endPos - overlapChars;
-    const lastChunkPos = chunks.at(-1)!.pos;
-    if (charPos <= lastChunkPos) {
-      // Prevent infinite loop - move forward at least a bit
-      charPos = endPos;
-    }
-  }
-
-  return chunks;
-}
-
-/**
- * Chunk a document by actual token count using the LLM tokenizer.
- * More accurate than character-based chunking but requires async.
- */
-export async function chunkDocumentByTokens(
-  content: string,
-  maxTokens: number = CHUNK_SIZE_TOKENS,
-  overlapTokens: number = CHUNK_OVERLAP_TOKENS,
-  windowTokens: number = CHUNK_WINDOW_TOKENS
-): Promise<{ text: string; pos: number; tokens: number }[]> {
-  const llm = getDefaultLLM();
-
-  // Use moderate chars/token estimate (prose ~4, code ~2, mixed ~3)
-  // If chunks exceed limit, they'll be re-split with actual ratio
-  const avgCharsPerToken = 3;
-  const maxChars = maxTokens * avgCharsPerToken;
-  const overlapChars = overlapTokens * avgCharsPerToken;
-  const windowChars = windowTokens * avgCharsPerToken;
-
-  // Chunk in character space with conservative estimate
-  let charChunks = chunkDocument(content, maxChars, overlapChars, windowChars);
-
-  // Tokenize and split any chunks that still exceed limit
-  const results: { text: string; pos: number; tokens: number }[] = [];
-
-  for (const chunk of charChunks) {
-    let tokensLength: number;
-    try {
-      if (llm.tokenize) {
-        // CJK and dense text can have very different chars-per-token ratios;
-        // tokenize() may throw if the text exceeds the model context window.
-        // Gracefully skip the chunk rather than crashing the entire embed run.
-        tokensLength = (await llm.tokenize(chunk.text)).length;
-      } else {
-        tokensLength = Math.ceil(chunk.text.length / 3.5);
-      }
-    } catch (tokenizeErr) {
-      process.stderr.write(
-        `KINDX Warning: tokenize() failed for chunk at pos=${chunk.pos} (len=${chunk.text.length}), skipping. ${tokenizeErr}\n`
-      );
-      // Apply a conservative character-based estimate and attempt to continue.
-      // If the chunk genuinely overflows the context, the embedding step will fail
-      // and the store.insertEmbedding call will never be reached.
-      tokensLength = Math.ceil(chunk.text.length / 2); // conservative: 2 chars/token for CJK
-    }
-
-    if (tokensLength <= maxTokens) {
-      results.push({ text: chunk.text, pos: chunk.pos, tokens: tokensLength });
-    } else {
-      // Chunk is still too large — split it further.
-      // Use actual char ratio to estimate a safe char limit with 5% safety margin.
-      const actualCharsPerToken = Math.max(1, chunk.text.length / tokensLength);
-      const safeMaxChars = Math.floor(maxTokens * actualCharsPerToken * 0.95);
-
-      const subChunks = chunkDocument(
-        chunk.text,
-        safeMaxChars,
-        Math.floor(overlapChars * actualCharsPerToken / 2),
-        Math.floor(windowChars * actualCharsPerToken / 2)
-      );
-
-      for (const subChunk of subChunks) {
-        let subTokensLength: number;
-        try {
-          if (llm.tokenize) {
-            subTokensLength = (await llm.tokenize(subChunk.text)).length;
-          } else {
-            subTokensLength = Math.ceil(subChunk.text.length / 3.5);
-          }
-        } catch {
-          // Sub-chunk still too large for tokenizer — use conservative estimate.
-          subTokensLength = Math.ceil(subChunk.text.length / 2);
-        }
-
-        // Only include sub-chunks that fit within the model context.
-        if (subTokensLength <= maxTokens) {
-          results.push({
-            text: subChunk.text,
-            pos: chunk.pos + subChunk.pos,
-            tokens: subTokensLength,
-          });
-        } else {
-          process.stderr.write(
-            `KINDX Warning: sub-chunk at pos=${chunk.pos + subChunk.pos} (tokens≈${subTokensLength}) exceeds maxTokens=${maxTokens}, skipping to prevent model overflow.\n`
-          );
-        }
-      }
-    }
-  }
-
-  return results;
-}
+// Chunking moved to engine/repository/chunking.ts (W1 C6).
 
 // =============================================================================
 // Fuzzy matching
